requirements.txt
streamlit
pandas
numpy
plotly
Pillow
sqlalchemy
werkzeug

streamlit-paste-button

audio-recorder-streamlit
google-genai
python-dotenv
